alicia-monitoring-app/backend/requirements.txt
# Alicia Monitoring App - Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# HTTP Client
aiohttp==3.9.1
httpx==0.25.2

# MQTT
paho-mqtt==1.6.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
